% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{clean_species_name}
\alias{clean_species_name}
\title{Clean binomial species names to retain only Genus + Species information.}
\usage{
clean_species_name(species, ignore_hybrids = FALSE)
}
\arguments{
\item{species}{A character string with binomial species name and potentially
additional information that want to be removed (see example)}

\item{ignore_hybrids}{If TRUE, the hybrids names are kept untouched. If
FALSE (the default) the hybrid names will be transformed either into
"Genus sp." or the first species of the hybrid name will be kept (which
is not necessarily what you want.)}
}
\value{
A character string with binomial names
}
\description{
The function will remove any text after (and including) a second space and
will try to transform as "Genus sp." any name that is ambiguous (e.g.
Canis cfr. lupus, Canis nr. lupus,...).
}
\examples{

example <- c(
"Acer tataricum subsp. aidzuense",
"Amphipappus fremontii var. spinosus",
"Helictochloa xenotus subsp. aff. pratensis GW-2014",
"Artocarpus nitidus cf. subsp. humilis EMG-2016",
"Artocarpus cf. nitidus",
"Artocarpus aff.nitidus",
"Artocarpus aff.",
"Artocarpus x",
"Artocarpus ",
"x ",
"Tilia x europaea",
"Tilia x-europaea",
"Salix x fragilis",
"Salix alba x S. fragilis",
"Salix viminalis x capreae",
"Messor minor x Messor cf. wasmanni",
"Messor sp. Pol &amp; Arnan 19645 (CBFS)",
"Crambidia nr. cephalica RDNMJ284-11",
"Lasiocampidae gen. lasioBioLep01 sp. BioLep01",
"Tomicus n. sp. CK-2004",
"Acronicta prob. iria NIBGE MOT-02051",
"Amata pr. huebneri",
"Quercus NA",
"Quercus x"
)

data.frame(input = example,
           output = clean_species_name(example),
           ignore_hybrids = clean_species_name(example, ignore_hybrids = TRUE)
)

}
