% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank_download.R
\name{genbank_count}
\alias{genbank_count}
\title{Number of sequences available on GenBank for a given Entrez query}
\usage{
genbank_count(entrez_query)
}
\arguments{
\item{entrez_query}{A character string containing the Entrez query.
It can contain spaces and new lines (which
are replaced internally by \code{+} signs to make the query string valid).}
}
\value{
An integer
}
\description{
Uses the NCBI API to retrieve the number of sequences available in GenBank
based on the provided Entrez query
}
\examples{

genbank_count("Adalia[organism] AND (COI[Title] or CO1[Title])")

}
