% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank_download.R
\name{genbank_download}
\alias{genbank_download}
\title{Downloads FASTA sequences and taxonomy information from NCBI GenBank database
based on a given Entrez query.}
\usage{
genbank_download(
  entrez_query,
  retmax = 350,
  seq_file_name = "sequences.fasta",
  taxo_file_name = "taxonomy.tsv",
  append = FALSE,
  verbose = TRUE,
  remove_tmp_files = TRUE
)
}
\arguments{
\item{entrez_query}{A character string containing the Entrez query used to
retrieve sequences from GenBank. It can contain spaces and new lines (which
are replaced internally by \code{+} signs to make the query string valid).}

\item{retmax}{An integer value indicating the maximum number of sequences
to retrieve per batch (default: 350). Higher values than 350 will
most likely cause systematically server errors. This values shoulp
not be changed in most of the cases.}

\item{seq_file_name}{A character string specifying the name of the output
FASTA file (default: "sequences.fasta").}

\item{taxo_file_name}{A character string specifying the name of the output
taxonomy file (default: "taxonomy.tsv").}

\item{append}{A logical value indicating whether to append the sequences
and taxonomy to existing files or create new ones (default: FALSE).
If FALSE, any existing file with the names indicated by \code{seq_file_name} and
\code{taxo_file_name} will be silently erased.}

\item{verbose}{A logical value indicating whether to display progress
messages and time elapsed (default: TRUE).}

\item{remove_tmp_files}{A logical value indicating whether to remove
2 temporary files (default: TRUE) : 1) \verb{__accver_taxid__.tsv} containing
the accession.version IDs and their corresponding taxid.
2) \verb{__taxonomy_long__.tsv}}
}
\value{
The function returns nothing but saves the results into 2 files
on the disk. The fasta file containing the sequences is progressively updated
at each batch. The taxonomy file is computed when all batches are completed.
}
\description{
The sequences and their taxonomy are downloaded by batches of size = \code{retmax}
and saved on the disk in two separated files.
In case of server error, a new request is sent repeatedly with increasing
waiting time. Typical usage : build a local reference database for barcoding
or metabarcoding.
}
\examples{

\dontrun{
# COI sequences for the genus Pullex
# Note that we can have spaces and new lines inside the query

my_query <-
"Pulex[Organism] AND
(CO1[Title] OR COI[Title] OR
cytochrome c oxidase subunit 1[Title] OR cytochrome c oxidase subunit I[Title] OR
cytochrome oxidase subunit 1[Title] OR cytochrome oxidase subunit I[Title]) AND
100:10000000[Sequence Length]) NOT
(uncultured OR environmental sample OR incertae sedis OR unverified)"

genbank_download(my_query)

# We can download the sequences of another flea genus and append the results
# to the existing files :

my_query <- "Xenopsylla[Organism] AND (CO1[Title] OR COI[Title])"
genbank_download(my_query, append = TRUE)

}

}
