% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{pivot_assign_taxonomy}
\alias{pivot_assign_taxonomy}
\title{Pivot outputs of \code{assign_taxonomy} from a long to a wide format}
\usage{
pivot_assign_taxonomy(assigned)
}
\arguments{
\item{assigned}{A data.frame produced by the function \code{\link[=assign_taxonomy]{assign_taxonomy()}}}
}
\value{
A data.frame with the same data as the input but reorganized in a wide format
}
\description{
This function will simply reorganize the output of the function \code{\link[=assign_taxonomy]{assign_taxonomy()}}
in order to obtain one row for each \code{Method} and each query \code{ID} with the
results of each taxonomic level side by side instead of being on top of each other.
For example the species level output will be placed in columns
\code{s_Taxon_true}, \code{s_Taxon}, \code{s_Ident},  \code{s_Cons}, \code{s_Length}, \code{s_Bit}.
The other taxonomic levels results will be placed in columns with similar names
with a different prefix : \code{g_} for genus, \code{f_} for family, \code{o_} for order,
\code{c_} for class, \code{p_} for phylum and \code{k_} for kingdom.
}
\examples{

# Fake top 15 blast hits for a sequence corresponding to Scirpoides holoschoenus
# and inspired by a real life example where the 4 assignment methods give
# different results

example_top15_blast_hits <-
    data.frame(
        TaxID_query = "ID001",
        Species_true = "Scirpoides holoschoenus",

        Bit_score = c(1075,704,682,651,630,630,549,549,529,525,525,520,520,518,512),
        Length = c(589,607,608,595,597,597,568,568,570,541,571,616,609,569,569),
        Identity = c(99.660,88.138,87.500,86.891,86.265,86.265,84.683,84.507,83.860,
                     84.843,83.713,82.630,82.594,83.480,83.304),
        E_value = 10^-14,

        Phylum = "Streptophyta",
        Class = "Magnoliopsida",
        Order = "Poales",
        Family = "Cyperaceae",
        Genus = c("Scirpoides","Erioscirpus","Erioscirpus","Dracoscirpoides",
                  "Dracoscirpoides","Dracoscirpoides", "Cyperus","Cyperus","Cyperus",
                  "Cyperus","Cyperus","Cyperus","Cyperus","Cyperus","Cyperus"),
        Species = c("Scirpoides holoschoenus","Erioscirpus microstachyus",
                    "Erioscirpus comosus", "Dracoscirpoides ficinioides",
                    "Dracoscirpoides falsa", "Dracoscirpoides falsa",
                    "Cyperus pulchellus", "Cyperus leucocephalus"," Cyperus haspan",
                    "Cyperus spiralis", "Cyperus flaccidus", "Cyperus sp.",
                    "Cyperus diffusus", "Cyperus tenuispica", "Cyperus haspan")
    )

# corresponding true taxonomy
example_taxonomy <-
data.frame(ID = "ID001",
Taxo = "k__Viridiplantae; p__Streptophyta; c__Magnoliopsida; o__Poales; f__Cyperaceae; g__Scirpoides; s__holoschoenus")


# Assign the taxonomy with mostly default options and in "long" format

tax_pred <-
    assign_taxonomy(df = example_top15_blast_hits,
                    taxo = example_taxonomy,
                    Order = NA, Family = NA) # ignore the family and genus levels
tax_pred


# Present the same data in a different, wider, format

pivot_assign_taxonomy(tax_pred)


}
